revert FlashAttention add more headdim
