+++ src/ATen/native/transformers/xpu/flash_attn/sycltla/mha_fwd.cpp
@@ -38,57 +38,315 @@
  */
 
 #include <sycltla/mha_common.h>
-#include <sycltla/mha_fwd_launch.h>
+#include <sycltla/mha_fwd.h>
+
+// batch, numhead_qo,numhead_kv,seqlen_qo,seqlen_kv,headsize_qk,headsize_vo
+using ProblemShapeRegular = cute::tuple<int, int, int, int, int, int, int>;
 
 namespace cute {
 
-// Explicit specializations for all headdim variants (merged into a single
-// compilation unit to reduce peak memory during parallel builds).
+template <class...>
+class MhaName;
 
-#define INSTANTIATE_FWD_HDIM(T, HDIM, CAUSAL, hdim_fn)  \
-  template <>                                           \
-  void run_mha_fwd_<T, HDIM, CAUSAL>(                   \
-      sycl::queue & queue, FLASH_FWD_params & params) { \
-    hdim_fn<T, CAUSAL>(queue, params);                  \
+template <class FMHAPrefillKernel, bool isVarLen>
+struct FA2Runner {
+  using ElementQ = typename FMHAPrefillKernel::ElementQ;
+  using ElementK = typename FMHAPrefillKernel::ElementK;
+  using ElementV = typename FMHAPrefillKernel::ElementV;
+  using ElementO = typename FMHAPrefillKernel::ElementO;
+
+  using ProblemShapeType = cutlass::fmha::kernel::FMHAProblemShape<isVarLen>;
+
+  //
+  // Methods
+  //
+
+  // Note that the GemmUniversalAdapter currently doesn't support flash
+  // attention, which is why this secondary `run` function is required to launch
+  // the kernel.
+  void run(sycl::queue& queue, typename FMHAPrefillKernel::Params params) {
+    namespace syclex = sycl::ext::oneapi::experimental;
+    namespace intelex = sycl::ext::intel::experimental;
+
+    dim3 const block = FMHAPrefillKernel::get_block_shape();
+    dim3 const grid = FMHAPrefillKernel::get_grid_shape(params);
+
+    // configure smem size and carveout
+    int smem_size = FMHAPrefillKernel::SharedStorageSize;
+
+    const auto sycl_block = compat::dim3(block.x, block.y, block.z);
+    const auto sycl_grid = compat::dim3(grid.x, grid.y, grid.z);
+
+    // Launch parameters depend on whether SYCL compiler supports work-group
+    // scratch memory extension
+    compat::experimental::launch_properties launch_props{
+        syclex::work_group_scratch_size(smem_size),
+    };
+    compat::experimental::kernel_properties kernel_props{
+        syclex::sub_group_size<cute::intel::sg_size>, intelex::grf_size<256>};
+    compat::experimental::launch_policy policy{
+        sycl_grid, sycl_block, launch_props, kernel_props};
+    compat::experimental::launch<
+        cutlass::device_kernel<FMHAPrefillKernel>,
+        MhaName<FMHAPrefillKernel>>(policy, queue, params);
   }
 
-#define INSTANTIATE_FWD_HDIM_ALL(HDIM, hdim_fn)                \
-  INSTANTIATE_FWD_HDIM(cute::half_t, HDIM, false, hdim_fn)     \
-  INSTANTIATE_FWD_HDIM(cute::half_t, HDIM, true, hdim_fn)      \
-  INSTANTIATE_FWD_HDIM(cute::bfloat16_t, HDIM, false, hdim_fn) \
-  INSTANTIATE_FWD_HDIM(cute::bfloat16_t, HDIM, true, hdim_fn)
+  void run(
+      sycl::queue& queue,
+      FLASH_FWD_params& params,
+      const cutlass::KernelHardwareInfo& hw_info) {
+    int batch = params.batch_size;
+    int num_heads_qo = params.num_heads_qo;
+    int num_heads_kv = params.num_heads_kv;
+    int seq_len_qo = params.seqlen_qo;
+    int seq_len_kv = params.seqlen_kv;
+    int head_size_qk = params.head_size_qk;
+    int head_size_vo = params.head_size_vo;
 
-INSTANTIATE_FWD_HDIM_ALL(32, run_mha_fwd_hdim32)
-INSTANTIATE_FWD_HDIM_ALL(64, run_mha_fwd_hdim64)
-INSTANTIATE_FWD_HDIM_ALL(96, run_mha_fwd_hdim96)
-INSTANTIATE_FWD_HDIM_ALL(128, run_mha_fwd_hdim128)
-INSTANTIATE_FWD_HDIM_ALL(192, run_mha_fwd_hdim192)
-INSTANTIATE_FWD_HDIM_ALL(256, run_mha_fwd_hdim256)
+    ProblemShapeType shape;
+    shape.batch = batch;
+    shape.num_heads_q = num_heads_qo;
+    shape.num_heads_kv = num_heads_kv;
+    shape.seq_len_qo = seq_len_qo;
+    shape.seq_len_kv = seq_len_kv;
+    shape.head_size_qk = head_size_qk;
+    shape.head_size_vo = head_size_vo;
 
-#undef INSTANTIATE_FWD_HDIM_ALL
-#undef INSTANTIATE_FWD_HDIM
+    const ElementQ* q_ptr = static_cast<const ElementQ*>(params.q_ptr);
+    const ElementK* k_ptr = static_cast<const ElementK*>(params.k_ptr);
+    const ElementV* v_ptr = static_cast<const ElementV*>(params.v_ptr);
+    ElementO* o_ptr = static_cast<ElementO*>(params.o_ptr);
+    float* lse_ptr = static_cast<float*>(params.lse_ptr);
+    float softmax_scale = params.scale;
+
+    typename FMHAPrefillKernel::Arguments arguments{
+        {
+            shape,
+            q_ptr,
+            params.q_batch_stride,
+            params.q_head_stride,
+            params.q_row_stride,
+            k_ptr,
+            params.k_batch_stride,
+            params.k_head_stride,
+            params.k_row_stride,
+            v_ptr,
+            params.v_batch_stride,
+            params.v_head_stride,
+            params.v_row_stride,
+            o_ptr,
+            params.o_batch_stride,
+            params.o_head_stride,
+            params.o_row_stride,
+            lse_ptr,
+        },
+        {softmax_scale},
+        {},
+        hw_info};
+
+    // Define device-global scratch memory
+    size_t workspace_size = FMHAPrefillKernel::get_workspace_size(arguments);
+    at::Tensor workspace_tensor = at::empty(
+        {static_cast<int64_t>(workspace_size)},
+        at::device(at::kXPU).dtype(at::kByte));
+
+    if (!FMHAPrefillKernel::can_implement(arguments)) {
+      TORCH_CHECK(
+          false,
+          "Invalid Problem Size",
+          batch,
+          "x",
+          num_heads_qo,
+          "x",
+          seq_len_qo,
+          "x",
+          seq_len_kv,
+          "x",
+          head_size_qk,
+          "x",
+          head_size_vo);
+      return;
+    }
+
+    // Initialize the workspace
+    CUTLASS_CHECK(FMHAPrefillKernel::initialize_workspace(
+        arguments, workspace_tensor.data_ptr()));
+
+    // Convert host-side arguments to device-side arguments to be passed to the
+    // kernel
+    auto kernel_params = FMHAPrefillKernel::to_underlying_arguments(
+        arguments, workspace_tensor.data_ptr());
+
+    // Launch a SYCL kernel using scratch/shared memory
+    run(queue, kernel_params);
+  }
+};
+
+template <
+    typename T,
+    bool IS_CAUSAL,
+    typename TileShapeQK,
+    typename TileShapePV,
+    typename TileShapeOutPut,
+    typename SubgroupLayoutQK,
+    int PipelineStages,
+    bool isVarLen = false>
+void run_mha_fwd_(sycl::queue& queue, FLASH_FWD_params& params) {
+  using ElementQ = T;
+  using ElementK = T;
+  using ElementV = T;
+  using ElementO = T;
+  using StrideQ = Stride<int64_t, _1, int64_t, int64_t>;
+  using StrideK = Stride<int64_t, _1, int64_t, int64_t>;
+  using StrideV = Stride<_1, int64_t, int64_t, int64_t>;
+  using StrideO = Stride<int64_t, _1, int64_t, int64_t>;
+  auto make_dummy_tensor = [&](auto val, auto stride) {
+    return make_tensor(
+        make_gmem_ptr(static_cast<decltype(val)*>(nullptr)),
+        make_layout(repeat<rank_v<decltype(stride)>>(1), stride));
+  };
+  auto make_const_dummy_tensor = [&](auto val, auto stride) {
+    return make_tensor(
+        make_gmem_ptr(static_cast<const decltype(val)*>(nullptr)),
+        make_layout(repeat<rank_v<decltype(stride)>>(1), stride));
+  };
+  using TensorQ = decltype(make_const_dummy_tensor(ElementQ{}, StrideQ{}));
+  using TensorK = decltype(make_const_dummy_tensor(ElementK{}, StrideK{}));
+  using TensorV = decltype(make_const_dummy_tensor(ElementV{}, StrideV{}));
+  using TensorO = decltype(make_dummy_tensor(ElementO{}, StrideO{}));
+
+  static constexpr int SGTileQ =
+      get<0>(shape_div(TileShapeQK{}, shape(SubgroupLayoutQK{})))();
+  static_assert(SGTileQ <= 16, "Subgroup tile in Q dimension must be <= 16");
+  using MMAOperation = XE_DPAS_TT<cute::gcd(SGTileQ, 8), float, T>;
+  using SubgroupLayoutPV =
+      decltype(cutlass::fmha::collective::get_sg_layout_pv(SubgroupLayoutQK{}));
+  using TiledMMAQK = typename TiledMMAHelper<
+      MMA_Atom<MMAOperation>,
+      Layout<TileShapeQK>,
+      SubgroupLayoutQK>::TiledMMA;
+  using TiledMMAPV = typename TiledMMAHelper<
+      MMA_Atom<MMAOperation>,
+      Layout<TileShapePV>,
+      SubgroupLayoutPV>::TiledMMA;
+  static_assert(
+      get<0>(TileShapeOutPut{}) == get<0>(TileShapePV{}),
+      "Output tile and P*V tile have different sizes in Q dimension");
+  constexpr int VTiles = get<1>(TileShapeOutPut{}) / get<1>(TileShapePV{});
+
+  cutlass::KernelHardwareInfo hw_info;
+
+  // Mainloop
+  using MainloopDispatchPolicy = cutlass::fmha::XeDefault<PipelineStages>;
+  using CollectiveMainloop = cutlass::fmha::collective::FMHAFwdMainloop<
+      MainloopDispatchPolicy,
+      IS_CAUSAL,
+      TiledMMAQK,
+      TiledMMAPV,
+      VTiles,
+      TensorQ,
+      TensorK,
+      TensorV>;
+
+  // Epilogue
+  using CollectiveEpilogue = cutlass::fmha::collective::
+      FMHAFwdEpilogue<CollectiveMainloop, TileShapeOutPut, TensorO>;
+
+  using Scheduler = cutlass::fmha::kernel::XeFMHAIndividualTileScheduler;
+  using ProblemShapeType = cutlass::fmha::kernel::FMHAProblemShape<isVarLen>;
+  using FMHAPrefillKernel = cutlass::fmha::kernel::XeFMHAFwdKernel<
+      ProblemShapeType,
+      CollectiveMainloop,
+      CollectiveEpilogue,
+      Scheduler>;
+
+  FA2Runner<FMHAPrefillKernel, isVarLen> runner;
+  runner.run(queue, params, hw_info);
+}
 
 template <typename T, bool IS_CAUSAL>
 void run_mha_fwd_(sycl::queue& queue, FLASH_FWD_params& params) {
   const int headdim = params.head_size_vo;
 
-  if (headdim <= 32) {
-    run_mha_fwd_<T, 32, IS_CAUSAL>(queue, params);
-  } else if (headdim <= 64) {
-    run_mha_fwd_<T, 64, IS_CAUSAL>(queue, params);
-  } else if (headdim <= 96) {
-    run_mha_fwd_<T, 96, IS_CAUSAL>(queue, params);
-  } else if (headdim <= 128) {
-    run_mha_fwd_<T, 128, IS_CAUSAL>(queue, params);
-  } else if (headdim <= 192) {
-    run_mha_fwd_<T, 192, IS_CAUSAL>(queue, params);
-  } else if (headdim <= 256) {
-    run_mha_fwd_<T, 256, IS_CAUSAL>(queue, params);
+#define run_mha_fwd_specialized( \
+    TileShapeQK_,                \
+    TileShapePV_,                \
+    TileShapeOutPut_,            \
+    SubgroupLayoutQK_,           \
+    PipelineStages_)             \
+  run_mha_fwd_<                  \
+      T,                         \
+      IS_CAUSAL,                 \
+      TileShapeQK_,              \
+      TileShapePV_,              \
+      TileShapeOutPut_,          \
+      SubgroupLayoutQK_,         \
+      PipelineStages_>(queue, params);
+
+  constexpr int PipelineStages = 2;
+  if (headdim == 64) {
+    int64_t batch_size = params.batch_size;
+    int64_t num_heads_qo = params.num_heads_qo;
+    int64_t seqlen_qo = params.seqlen_qo;
+    if (batch_size * num_heads_qo * seqlen_qo <= 8192) {
+      using TileShapeQK = Shape<_64, _64, _32>;
+      using TileShapePV = Shape<_64, _32, _64>;
+      using TileShapeOutPut = Shape<_64, _64>;
+      using SubgroupLayoutQK = Layout<Shape<_16, _1, _1>>;
+      run_mha_fwd_specialized(
+          TileShapeQK,
+          TileShapePV,
+          TileShapeOutPut,
+          SubgroupLayoutQK,
+          PipelineStages);
+    } else {
+      using TileShapeQK = Shape<_128, _64, _32>;
+      using TileShapePV = Shape<_128, _32, _64>;
+      using TileShapeOutPut = Shape<_128, _64>;
+      using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
+      run_mha_fwd_specialized(
+          TileShapeQK,
+          TileShapePV,
+          TileShapeOutPut,
+          SubgroupLayoutQK,
+          PipelineStages);
+    }
+  } else if (headdim == 96) {
+    using TileShapeQK = Shape<_128, _64, _32>;
+    using TileShapePV = Shape<_128, _32, _64>;
+    using TileShapeOutPut = Shape<_128, _96>;
+    using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
+    run_mha_fwd_specialized(
+        TileShapeQK,
+        TileShapePV,
+        TileShapeOutPut,
+        SubgroupLayoutQK,
+        PipelineStages);
+  } else if (headdim == 128) {
+    using TileShapeQK = Shape<_128, _32, _32>;
+    using TileShapePV = Shape<_128, _32, _32>;
+    using TileShapeOutPut = Shape<_128, _128>;
+    using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
+    run_mha_fwd_specialized(
+        TileShapeQK,
+        TileShapePV,
+        TileShapeOutPut,
+        SubgroupLayoutQK,
+        PipelineStages);
+  } else if (headdim == 192) {
+    using TileShapeQK = Shape<_256, _64, _32>;
+    using TileShapePV = Shape<_256, _32, _64>;
+    using TileShapeOutPut = Shape<_256, _192>;
+    using SubgroupLayoutQK = Layout<Shape<_32, _1, _1>>;
+    run_mha_fwd_specialized(
+        TileShapeQK,
+        TileShapePV,
+        TileShapeOutPut,
+        SubgroupLayoutQK,
+        PipelineStages);
   } else {
     TORCH_CHECK(
-        false,
-        "FlashAttentionForwardXPU only support headdim up to 256, got ",
-        headdim);
+        false, "FlashAttentionForwardXPU only support headdim 64,96,128,192");
   }
 }
 
@@ -171,44 +429,19 @@ flash_attention_forward_sycltla(
   TORCH_CHECK(
       value.stride(-1) == 1,
       "FlashAttentionForwardXPU: input tensor must have contiguous last dimension");
-  TORCH_CHECK(
-      headsize_vo == headsize_qk,
-      "FlashAttentionForwardXPU only support headsize_qk equal to headsize_vo");
-
-  const int headsize_padded = round_up_headdim(headsize_qk);
-  const bool needs_headsize_pad = headsize_padded > headsize_qk;
-
-  at::Tensor q_padded = query;
-  at::Tensor k_padded = key;
-  at::Tensor v_padded = value;
-  if (needs_headsize_pad) {
-    int pad = headsize_padded - headsize_qk;
-    q_padded = at::constant_pad_nd(query, {0, pad}, 0);
-    k_padded = at::constant_pad_nd(key, {0, pad}, 0);
-    v_padded = at::constant_pad_nd(value, {0, pad}, 0);
-  }
 
   auto opts = query.options();
   at::Tensor out = at::empty_like(query);
-
-  at::Tensor out_padded = needs_headsize_pad
-      ? at::empty({batch_size, numhead_qo, seqlen_qo, headsize_padded}, opts)
-      : out;
 
   if (seqlen_qo > seqlen_kv && is_causal) {
     // When seqlen_qo is greater than seqlen_kv and is_causal(lower_right causal
     // mask) is true, some output positions will skip computation for better
     // performance.
-    out_padded.zero_();
+    out.zero_();
   }
 
   at::Tensor logsumexp =
       at::empty({batch_size, numhead_qo, seqlen_qo}, opts.dtype(at::kFloat));
-
-  // Base pointers must be 64-byte aligned for block 2D load
-  at::Tensor q_aligned = ensure_alignment_for_sdpa(q_padded);
-  at::Tensor k_aligned = ensure_alignment_for_sdpa(k_padded);
-  at::Tensor v_aligned = ensure_alignment_for_sdpa(v_padded);
 
   auto sycl_queue = at::xpu::getCurrentXPUStream().queue();
   auto device_architecture =
@@ -238,23 +471,19 @@ flash_attention_forward_sycltla(
       numhead_kv,
       seqlen_qo,
       seqlen_kv,
-      headsize_padded,
-      headsize_padded,
+      headsize_qk,
+      headsize_vo,
       0, // unused seqlen_qo_pad
       0, // unused seqlen_kv_pad
-      q_aligned,
-      k_aligned,
-      v_aligned,
-      out_padded,
+      query,
+      key,
+      value,
+      out,
       logsumexp,
       scale,
       is_causal);
 
   cute::run_mha_fwd(sycl_queue, params);
-
-  if (needs_headsize_pad) {
-    out.copy_(out_padded.slice(/*dim=*/3, /*start=*/0, /*end=*/headsize_qk));
-  }
 
   return std::tuple<
       at::Tensor,
