--- src/ATen/native/transformers/xpu/flash_attn/sycltla/mha_fwd_launch.h
@@ -1,383 +1,0 @@
-/*
- * Copyright 2020-2026 Intel Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Copyright (c) 2024 - 2025 Codeplay Software Ltd. All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-// This file contains the template definitions for per-headdim forward kernels.
-// All explicit specializations are instantiated in mha_fwd.cpp.
-
-#pragma once
-
-#include <sycltla/mha_common.h>
-#include <sycltla/mha_fwd.h>
-
-// batch, numhead_qo,numhead_kv,seqlen_qo,seqlen_kv,headsize_qk,headsize_vo
-using ProblemShapeRegular = cute::tuple<int, int, int, int, int, int, int>;
-
-namespace cute {
-
-template <class...>
-class MhaName;
-
-template <class FMHAPrefillKernel, bool isVarLen>
-struct FA2Runner {
-  using ElementQ = typename FMHAPrefillKernel::ElementQ;
-  using ElementK = typename FMHAPrefillKernel::ElementK;
-  using ElementV = typename FMHAPrefillKernel::ElementV;
-  using ElementO = typename FMHAPrefillKernel::ElementO;
-
-  using ProblemShapeType = cutlass::fmha::kernel::FMHAProblemShape<isVarLen>;
-
-  void run(sycl::queue& queue, typename FMHAPrefillKernel::Params params) {
-    namespace syclex = sycl::ext::oneapi::experimental;
-    namespace intelex = sycl::ext::intel::experimental;
-
-    dim3 const block = FMHAPrefillKernel::get_block_shape();
-    dim3 const grid = FMHAPrefillKernel::get_grid_shape(params);
-
-    int smem_size = FMHAPrefillKernel::SharedStorageSize;
-
-    const auto sycl_block = compat::dim3(block.x, block.y, block.z);
-    const auto sycl_grid = compat::dim3(grid.x, grid.y, grid.z);
-
-    compat::experimental::launch_properties launch_props{
-        syclex::work_group_scratch_size(smem_size),
-    };
-    compat::experimental::kernel_properties kernel_props{
-        syclex::sub_group_size<cute::intel::sg_size>, intelex::grf_size<256>};
-    compat::experimental::launch_policy policy{
-        sycl_grid, sycl_block, launch_props, kernel_props};
-    compat::experimental::launch<
-        cutlass::device_kernel<FMHAPrefillKernel>,
-        MhaName<FMHAPrefillKernel>>(policy, queue, params);
-  }
-
-  void run(
-      sycl::queue& queue,
-      FLASH_FWD_params& params,
-      const cutlass::KernelHardwareInfo& hw_info) {
-    int batch = params.batch_size;
-    int num_heads_qo = params.num_heads_qo;
-    int num_heads_kv = params.num_heads_kv;
-    int seq_len_qo = params.seqlen_qo;
-    int seq_len_kv = params.seqlen_kv;
-    int head_size_qk = params.head_size_qk;
-    int head_size_vo = params.head_size_vo;
-
-    ProblemShapeType shape;
-    shape.batch = batch;
-    shape.num_heads_q = num_heads_qo;
-    shape.num_heads_kv = num_heads_kv;
-    shape.seq_len_qo = seq_len_qo;
-    shape.seq_len_kv = seq_len_kv;
-    shape.head_size_qk = head_size_qk;
-    shape.head_size_vo = head_size_vo;
-
-    const ElementQ* q_ptr = static_cast<const ElementQ*>(params.q_ptr);
-    const ElementK* k_ptr = static_cast<const ElementK*>(params.k_ptr);
-    const ElementV* v_ptr = static_cast<const ElementV*>(params.v_ptr);
-    ElementO* o_ptr = static_cast<ElementO*>(params.o_ptr);
-    float* lse_ptr = static_cast<float*>(params.lse_ptr);
-    float softmax_scale = params.scale;
-
-    typename FMHAPrefillKernel::Arguments arguments{
-        {
-            shape,
-            q_ptr,
-            params.q_batch_stride,
-            params.q_head_stride,
-            params.q_row_stride,
-            k_ptr,
-            params.k_batch_stride,
-            params.k_head_stride,
-            params.k_row_stride,
-            v_ptr,
-            params.v_batch_stride,
-            params.v_head_stride,
-            params.v_row_stride,
-            o_ptr,
-            params.o_batch_stride,
-            params.o_head_stride,
-            params.o_row_stride,
-            lse_ptr,
-        },
-        {softmax_scale},
-        {},
-        hw_info};
-
-    size_t workspace_size = FMHAPrefillKernel::get_workspace_size(arguments);
-    at::Tensor workspace_tensor = at::empty(
-        {static_cast<int64_t>(workspace_size)},
-        at::device(at::kXPU).dtype(at::kByte));
-
-    if (!FMHAPrefillKernel::can_implement(arguments)) {
-      TORCH_CHECK(
-          false,
-          "Invalid Problem Size",
-          batch,
-          "x",
-          num_heads_qo,
-          "x",
-          seq_len_qo,
-          "x",
-          seq_len_kv,
-          "x",
-          head_size_qk,
-          "x",
-          head_size_vo);
-      return;
-    }
-
-    CUTLASS_CHECK(FMHAPrefillKernel::initialize_workspace(
-        arguments, workspace_tensor.data_ptr()));
-
-    auto kernel_params = FMHAPrefillKernel::to_underlying_arguments(
-        arguments, workspace_tensor.data_ptr());
-
-    run(queue, kernel_params);
-  }
-};
-
-template <
-    typename T,
-    bool IS_CAUSAL,
-    typename TileShapeQK,
-    typename TileShapePV,
-    typename TileShapeOutPut,
-    typename SubgroupLayoutQK,
-    int PipelineStages,
-    bool isVarLen = false>
-void run_mha_fwd_impl(sycl::queue& queue, FLASH_FWD_params& params) {
-  using ElementQ = T;
-  using ElementK = T;
-  using ElementV = T;
-  using ElementO = T;
-  using StrideQ = Stride<int64_t, _1, int64_t, int64_t>;
-  using StrideK = Stride<int64_t, _1, int64_t, int64_t>;
-  using StrideV = Stride<_1, int64_t, int64_t, int64_t>;
-  using StrideO = Stride<int64_t, _1, int64_t, int64_t>;
-  auto make_dummy_tensor = [&](auto val, auto stride) {
-    return make_tensor(
-        make_gmem_ptr(static_cast<decltype(val)*>(nullptr)),
-        make_layout(repeat<rank_v<decltype(stride)>>(1), stride));
-  };
-  auto make_const_dummy_tensor = [&](auto val, auto stride) {
-    return make_tensor(
-        make_gmem_ptr(static_cast<const decltype(val)*>(nullptr)),
-        make_layout(repeat<rank_v<decltype(stride)>>(1), stride));
-  };
-  using TensorQ = decltype(make_const_dummy_tensor(ElementQ{}, StrideQ{}));
-  using TensorK = decltype(make_const_dummy_tensor(ElementK{}, StrideK{}));
-  using TensorV = decltype(make_const_dummy_tensor(ElementV{}, StrideV{}));
-  using TensorO = decltype(make_dummy_tensor(ElementO{}, StrideO{}));
-
-  static constexpr int SGTileQ =
-      get<0>(shape_div(TileShapeQK{}, shape(SubgroupLayoutQK{})))();
-  static_assert(SGTileQ <= 16, "Subgroup tile in Q dimension must be <= 16");
-  using MMAOperation = XE_DPAS_TT<cute::gcd(SGTileQ, 8), float, T>;
-  using SubgroupLayoutPV =
-      decltype(cutlass::fmha::collective::get_sg_layout_pv(SubgroupLayoutQK{}));
-  using TiledMMAQK = typename TiledMMAHelper<
-      MMA_Atom<MMAOperation>,
-      Layout<TileShapeQK>,
-      SubgroupLayoutQK>::TiledMMA;
-  using TiledMMAPV = typename TiledMMAHelper<
-      MMA_Atom<MMAOperation>,
-      Layout<TileShapePV>,
-      SubgroupLayoutPV>::TiledMMA;
-  static_assert(
-      get<0>(TileShapeOutPut{}) == get<0>(TileShapePV{}),
-      "Output tile and P*V tile have different sizes in Q dimension");
-  constexpr int VTiles = get<1>(TileShapeOutPut{}) / get<1>(TileShapePV{});
-
-  cutlass::KernelHardwareInfo hw_info;
-
-  using MainloopDispatchPolicy = cutlass::fmha::XeDefault<PipelineStages>;
-  using CollectiveMainloop = cutlass::fmha::collective::FMHAFwdMainloop<
-      MainloopDispatchPolicy,
-      IS_CAUSAL,
-      TiledMMAQK,
-      TiledMMAPV,
-      VTiles,
-      TensorQ,
-      TensorK,
-      TensorV>;
-
-  using CollectiveEpilogue = cutlass::fmha::collective::
-      FMHAFwdEpilogue<CollectiveMainloop, TileShapeOutPut, TensorO>;
-
-  using Scheduler = cutlass::fmha::kernel::XeFMHAIndividualTileScheduler;
-  using ProblemShapeType = cutlass::fmha::kernel::FMHAProblemShape<isVarLen>;
-  using FMHAPrefillKernel = cutlass::fmha::kernel::XeFMHAFwdKernel<
-      ProblemShapeType,
-      CollectiveMainloop,
-      CollectiveEpilogue,
-      Scheduler>;
-
-  FA2Runner<FMHAPrefillKernel, isVarLen> runner;
-  runner.run(queue, params, hw_info);
-}
-
-// Convenience macro for invoking run_mha_fwd_impl with tile shape parameters
-#define run_mha_fwd_specialized( \
-    TileShapeQK_,                \
-    TileShapePV_,                \
-    TileShapeOutPut_,            \
-    SubgroupLayoutQK_,           \
-    PipelineStages_)             \
-  run_mha_fwd_impl<              \
-      T,                         \
-      IS_CAUSAL,                 \
-      TileShapeQK_,              \
-      TileShapePV_,              \
-      TileShapeOutPut_,          \
-      SubgroupLayoutQK_,         \
-      PipelineStages_>(queue, params)
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim32(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  using TileShapeQK = Shape<_128, _64, _32>;
-  using TileShapePV = Shape<_128, _32, _64>;
-  using TileShapeOutPut = Shape<_128, _32>;
-  using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
-  run_mha_fwd_specialized(
-      TileShapeQK,
-      TileShapePV,
-      TileShapeOutPut,
-      SubgroupLayoutQK,
-      PipelineStages);
-}
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim64(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  int64_t batch_size = params.batch_size;
-  int64_t num_heads_qo = params.num_heads_qo;
-  int64_t seqlen_qo = params.seqlen_qo;
-  if (batch_size * num_heads_qo * seqlen_qo <= 8192) {
-    using TileShapeQK = Shape<_64, _64, _32>;
-    using TileShapePV = Shape<_64, _32, _64>;
-    using TileShapeOutPut = Shape<_64, _64>;
-    using SubgroupLayoutQK = Layout<Shape<_16, _1, _1>>;
-    run_mha_fwd_specialized(
-        TileShapeQK,
-        TileShapePV,
-        TileShapeOutPut,
-        SubgroupLayoutQK,
-        PipelineStages);
-  } else {
-    using TileShapeQK = Shape<_128, _64, _32>;
-    using TileShapePV = Shape<_128, _32, _64>;
-    using TileShapeOutPut = Shape<_128, _64>;
-    using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
-    run_mha_fwd_specialized(
-        TileShapeQK,
-        TileShapePV,
-        TileShapeOutPut,
-        SubgroupLayoutQK,
-        PipelineStages);
-  }
-}
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim96(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  using TileShapeQK = Shape<_128, _64, _32>;
-  using TileShapePV = Shape<_128, _32, _64>;
-  using TileShapeOutPut = Shape<_128, _96>;
-  using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
-  run_mha_fwd_specialized(
-      TileShapeQK,
-      TileShapePV,
-      TileShapeOutPut,
-      SubgroupLayoutQK,
-      PipelineStages);
-}
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim128(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  using TileShapeQK = Shape<_128, _32, _32>;
-  using TileShapePV = Shape<_128, _32, _32>;
-  using TileShapeOutPut = Shape<_128, _128>;
-  using SubgroupLayoutQK = Layout<Shape<_8, _1, _1>>;
-  run_mha_fwd_specialized(
-      TileShapeQK,
-      TileShapePV,
-      TileShapeOutPut,
-      SubgroupLayoutQK,
-      PipelineStages);
-}
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim192(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  using TileShapeQK = Shape<_256, _64, _32>;
-  using TileShapePV = Shape<_256, _32, _64>;
-  using TileShapeOutPut = Shape<_256, _192>;
-  using SubgroupLayoutQK = Layout<Shape<_32, _1, _1>>;
-  run_mha_fwd_specialized(
-      TileShapeQK,
-      TileShapePV,
-      TileShapeOutPut,
-      SubgroupLayoutQK,
-      PipelineStages);
-}
-
-template <typename T, bool IS_CAUSAL>
-void run_mha_fwd_hdim256(sycl::queue& queue, FLASH_FWD_params& params) {
-  constexpr int PipelineStages = 2;
-  using TileShapeQK = Shape<_256, _64, _32>;
-  using TileShapePV = Shape<_256, _32, _64>;
-  using TileShapeOutPut = Shape<_256, _256>;
-  using SubgroupLayoutQK = Layout<Shape<_32, _1, _1>>;
-  run_mha_fwd_specialized(
-      TileShapeQK,
-      TileShapePV,
-      TileShapeOutPut,
-      SubgroupLayoutQK,
-      PipelineStages);
-}
-
-// Primary template declaration -- explicit specializations are in mha_fwd.cpp.
-template <typename T, int Headdim, bool IS_CAUSAL>
-void run_mha_fwd_(sycl::queue& queue, FLASH_FWD_params& params);
-
-#undef run_mha_fwd_specialized
-
-} // namespace cute
